requirements.txt
django
djangorestframework
pillow
psycopg2-binary
python-slugify
python-decouple
django-filter
